A new method of using processes and a revised README file
