+++ config.env
@@ -6,8 +6,8 @@ CSVFile = nodes.csv
 firstContractTX = 0x90eeb0a0680034c8c340dfa60b773ed77b060d6a966596059610981486d312f4
 
 [DATABASE]
-engine = MYSQL
-# engine = SQLITE
+# engine = MYSQL
+engine = SQLITE
 queryLogIsEnabled = False
 
 [SQLITE]
