+++ config.py
@@ -1,40 +1,20 @@
 
 import configparser, os
-import logging.handlers
-from src.custom_formatter import CustomFormatter
-from datetime import datetime
-from multiprocessing import cpu_count
 
 config = configparser.ConfigParser()
-config.read('config.env')
+try:
+    config.read('config.env')
+except ImportError:
+    config.read(os.path.join(os.getcwd(), 'config.env'))
 
-from src.exceptions import DatabaseEngineNotFoundError
+from libs.exceptions import DatabaseEngineNotFoundError
 
 try:
     if config['DATABASE']['engine'] != 'MYSQL':
         raise DatabaseEngineNotFoundError('For DB Engine is used Sqlite')
-    from services.mysql_database import MysqlDatabase as Database, dbException, Singleton, DB_TYPES
+    from services.mysql_database import MysqlDatabase as Database, DB_TYPES, Singleton
     Database.ENGINE = DB_TYPES.MYSQL
-except (ImportError, DatabaseEngineNotFoundError) as e:
-    from services.sqlite_database import SqliteDatabase as Database, dbException, Singleton, DB_TYPES
+except (ImportError, DatabaseEngineNotFoundError):
+    from services.sqlite_database import SqliteDatabase as Database, DB_TYPES, Singleton
     Database.ENGINE = DB_TYPES.SQLITE
     print(Database.ENGINE)
-
-
-
-
-BASE_LOOP_ITER = 1000 
-LIMIT_OF_THREADS = int(cpu_count() * 8)
-FREE_MEMORY_IN_INTERVAL = True
-
-
-def getLogger():
-    logger = logging.getLogger("ETNY NODE")
-    handler = logging.handlers.RotatingFileHandler(os.path.join(os.getcwd(), f"output_{datetime.now().strftime('%d-%m-%Y')}.log"), maxBytes=20480000, backupCount=5)
-    fmt = '%(asctime)s %(message)s'
-    formatter = logging.Formatter(fmt)
-    handler.setFormatter(formatter)
-    # handler.setFormatter(CustomFormatter(fmt))
-    logger.addHandler(handler)
-    logger.setLevel(logging.INFO)
-    return logger
