+++ libs/exceptions.py
