+++ services/sqlite_database.py
@@ -1,56 +1,25 @@
-from socket import timeout
-from datetime import datetime
 import sqlite3, configparser
-from src.singleton import Singleton
-from services.database import Database, time, DB_TYPES
-from sqlite3 import OperationalError as dbException
+try:
+    from services.database import Singleton, Database, DB_TYPES
+except ImportError:
+    from database import Singleton, Database, DB_TYPES
 
 
 class SqliteDatabase(Database, metaclass = Singleton):
 
-    def reConnect(self, config = None) -> None:
-        self.connect(config=config)
-
-    def connect(self, config = None, has_dict_cursor = False):
+    def connect(self, config = None):
         super().connect(config=config)
         if not config:
             config = configparser.ConfigParser()
             config.read('config.env')
 
         self._conn = sqlite3.connect(config['SQLITE']['DB_DATABASE'], timeout=15)
-
-        if self.dict_cursor == True or has_dict_cursor:
-            self._conn.row_factory = sqlite3.Row
         self._curr = self._conn.cursor()
         
-
     def init(self, *args, **kwargs):
         Database.init(self, *args, **kwargs)
-        self._curr.execute(f'''CREATE TABLE IF NOT EXISTS {self.TABLE_NAME} (
-                id bigint(20) primary key, 
-                block_identifier bigint(20) default 0,
-                created_on int(11) not null default 0,
-                last_updated int(11) not null default 0,
-                updates_count int(11) not null default 0
-            )''')
-        self._curr.execute(f'''create table if not exists {self.TABLE_NAME}_details (
-                id bigint(20) primary key,
-                parent_id bigint(20) not null default 0,  
-                address varchar(70) not null default '', 
-                cpu tinyint(3) not null default 0, 
-                memory tinyint(3) not null default 0, 
-                storage tinyint(3) not null default 0, 
-                bandwith tinyint(3) not null default 0, 
-                duration tinyint(3) not null default 0, 
-                insert_date int(11) not null default 0,
-                update_date int(11) not null default 0,
-                status tinyint(3) not null default 0, 
-                cost decimal(10, 2) not null default 0.00,
-                FOREIGN KEY (parent_id) REFERENCES {self.TABLE_NAME}(id)
-            )''')
-
-        # self._curr.execute("drop table if exists dp_requests;")
-        self._curr.execute(f'''create table if not exists dp_requests (
+        
+        self._curr.execute('''create table if not exists dp_requests (
                 id bigint(20) primary key,
                 dpRequestId bigint(20) UNIQUE default 0,
                 dproc varchar(70) not null default '',
@@ -66,66 +35,69 @@ class SqliteDatabase(Database, metaclass = Singleton):
                 local_created_at int(11) not null default (cast(strftime('%s','now') as int))
             )''')
 
-        self._curr.execute(f'''CREATE INDEX if not exists parent_id ON {self.TABLE_NAME}_details (parent_id);''')
-        self._curr.execute(f'''CREATE INDEX if not exists address ON {self.TABLE_NAME}_details (address);''')
+        self._curr.execute('''create table if not exists dp_unique_requests (
+            id bigint(20) not null default 0,
+            dpRequestId bigint(20) UNIQUE default 0,
+            dproc varchar(70) unique not null default '',
+            cpuRequest tinyint not null default 0,
+            memoryRequest tinyint not null default 0,
+            storageRequest tinyint not null default 0,
+            bandwidthRequest tinyint not null default 0,
+            duration tinyint not null default 0,
+            minPrice tinyint not null default 0,
+            `status` tinyint not null default 0,
+            createdAt int(11) not null default 0,
+            local_created_at int(11) not null default (cast(strftime('%s','now') as int)),
+            updated_at int(11) not null default 0,
+            nodes_count mediumint default 0
+        )''')
 
-        self._curr.execute(f'''CREATE INDEX if not exists dp_requests_dproc_idx ON  dp_requests (dproc);''')
-        self._curr.execute(f'''CREATE INDEX if not exists dp_requests_dpRequestId_idx ON  dp_requests (dpRequestId);''')
-        self._curr.execute(f'''CREATE INDEX if not exists dp_requests_createdAt_idx ON  dp_requests (createdAt);''')
+        self._curr.execute('''CREATE INDEX if not exists dp_requests_dproc_idx ON  dp_requests (dproc);''')
+        self._curr.execute('''CREATE INDEX if not exists dp_requests_dpRequestId_idx ON  dp_requests (dpRequestId);''')
+        self._curr.execute('''CREATE INDEX if not exists dp_requests_createdAt_idx ON  dp_requests (createdAt);''')
+
+        self._curr.execute('''CREATE INDEX if not exists dp_unique_requests_dproc_idx ON  dp_unique_requests (dproc);''')
+        self._curr.execute('''CREATE INDEX if not exists dp_unique_requests_dpRequestId_idx ON  dp_unique_requests (dpRequestId);''')
+        self._curr.execute('''CREATE INDEX if not exists dp_unique_requests_createdAt_idx ON  dp_unique_requests (createdAt);''')
+        self._curr.execute('''CREATE INDEX if not exists dp_unique_requests_updated_at_idx ON  dp_unique_requests (updated_at);''')
+        
         self._conn.commit()
 
-    def insert(self, node, recursion_count = 0):
-        try:
-            public = node.public()
-            private = node.private()
-            # root
-            query = f'''insert into {self.TABLE_NAME} ({",".join(private.keys())}) values ({",".join([f"'{x}'" if type(x) == str else f'{x}' for x in [*private.values()]])})'''
-            
-            self._curr.execute(query)
-            insert_id = self._curr.lastrowid
-            if not insert_id:
-                raise Exception('insert id not found...')
-            # child
-            if self.queryLogIsEnabled:
-                print(query)
-            sub_query = f'''insert into {self.TABLE_NAME}_details ({",".join(public.keys())}, parent_id, insert_date) 
-                            values 
-                        (  {",".join([f"'{x}'" if type(x) == str else f'{x}' for x in [*public.values()]])}, {insert_id}, {int(time.time())} )'''
-            return self._curr.execute(sub_query)
-        except sqlite3.IntegrityError as e:
-            pass
-        except (self._conn.Error, Exception) as e:
-            if self.logger:
-                self.logger.error(f'error = {e}')
-            time.sleep(1)
-            if recursion_count < 10:
-                self.connect()
-                return self.insert(node=node, recursion_count=recursion_count+1)
-        finally:
-            self._conn.commit()
+    def get_missing_records_count(self):
+        self._curr.execute('select max(id) - count(id) from dp_requests')
+        result = self._curr.fetchone()
+        return result[0] if result else 0
 
-    def select(self):
-        for row in self._curr.execute(f"select * from {self.TABLE_NAME}"):
-            print(row)
+    def __get_dp_request_by_id(self, _id):
+        self._curr.execute(f'SELECT id FROM dp_requests where id = {_id}')
+        result = self._curr.fetchone()
+        return result[0] if result else 0
 
-    def commit(self) -> None:
-        try:
-            return super().commit()
-        except dbException as e:
-            if self.logger:
-                self.logger.error(f'db_error: {e}')
-            time.sleep(2)
-            self.connect()
-            return self.commit()
+    def get_missing_records(self, last_page = 1, per_page = 10):
+        _max = self.get_last_dp_request(field='id')
+        _str = ''
+        inline_query = 0
+        fount_items = 0
+        current_iter = last_page + 2 if last_page > 1 else last_page
+        while current_iter < _max and fount_items < per_page:
+            inline_query = self.__get_dp_request_by_id(_id = current_iter)
+            if not inline_query:
+                _str += f'{current_iter - 1},'
+                fount_items += 1
+            current_iter += 1
+        return f'{fount_items}-{current_iter}-{_max}-{_str}'
 
-    def select_all(self, limit = 1000):
-        query = super().select_all(limit=limit)
-        self._curr.execute(query)
-        result = self._curr.fetchall()
-        return self.generator_dict_from_result(result)
+    def generate_unique_requests(self):
+        sql = '''
+            insert or replace into dp_unique_requests 
+                ( id, dpRequestId, dproc, cpuRequest, memoryRequest, storageRequest, bandwidthRequest, duration, minPrice, `status`, createdAt, local_created_at, updated_at, nodes_count ) 
+                select max(id) as id, max(dpRequestId) as dpRequestId, dproc, cpuRequest, memoryRequest, storageRequest, bandwidthRequest, duration, minPrice, max(`status`) as `status`, createdAt, local_created_at, max(createdAt) as updated_at, count(id) as nodes_count
+            from dp_requests group by dproc
+        '''
+        self._curr.execute(sql)
+        self._conn.commit()
 
-    def generator_dict_from_result(self, result):
-        return ({'_id': result[i]['id'], **dict(zip(row.keys(), row))} for i, row in enumerate(result))
-        
-    def get_concatenated_fields(self):
-        return "(select block_identifier || '-' || (block_identifier - d.block_identifier) as id_and_block from orders where block_identifier > d.block_identifier order by block_identifier asc limit 1) as next_id_and_block_identifier"
+    def get_unique_requests_count(self, interval_hours = 24):
+        super().get_count_of_dp_requests()
+        self._curr.execute(f"select count(distinct dproc) from dp_requests where createdAt >= cast(strftime('%s', datetime('now', '-{interval_hours} hour')) as integer)")
+        return self._curr.fetchone()[0]
