+++ sqlite_requirements.txt
@@ -1,0 +1,2 @@
+web3==5.31.3
+packaging==20.3
