--- src/exceptions.py
@@ -1,7 +1,0 @@
-class NotFoundException(Exception):
-    pass
-class DatabaseEngineNotFoundError(Exception):
-    pass
-
-class BreackFromLoopException(Exception):
-    pass
